refactor: dto 패키지 service 내부로 옮김
